Fix the stobj drawing issue with high itemgroup counts!
Stage objects (party balls, goaltape, etc) reference the itemgroup they're on as a
signed byte, so on a stage with over 127 itemgroups it'll be treated as negative.
So indexing the game's list of current itemgroup animations was done with negative values,
giving garbage values for the transforms used to draw stobjs.
I've patched all the places I could (easily) find that treat it as signed
so the limit should be 255 itemgroups now.